requirements.txt
colorama>=0.4.4
Pillow>=8.4.0
anytree>=2.8.0